feat: setup project root routes & error page
