+++ react-client/src/routes/AppRoutePaths.ts
@@ -1,0 +1,5 @@
+export const AppRoutePaths = {
+  root: "/",
+  account: "account",
+  myPets: "my-pets",
+};
